Put classes in learning, and add unit tests.
